Reorganize the NonlinearSystem hierarchy
to allow us to derive EigenSystem from Libmesh. To achieve this goal, the following steps will be taken:

(1) Rename  NonlinearSystem to NonlinearSystemBase

(2) Create a new class NonlinearSystem deriving from the class NonlinearSystemBase

(3) Change the inheritance of the class MooseEigenSystem. MooseEigenSystem should  derive  from the class NonlinearSystemBase instead of the class NonlinearSystem

(4) Move some functions specified for the NonlinearSystem only from the NonlinearSystemBase  to the class NonlinearSystem

(5) Enable SystemTempl<EigenSystem> to bring SLEPc from Libmesh

In this PR, we finished steps (1) and (2).

Refs  #7398
