+++ framework/include/base/ComputeDiracThread.h
@@ -23,7 +23,7 @@
 #include "libmesh/elem_range.h"
 
 // Forward declarations
-class NonlinearSystem;
+class NonlinearSystemBase;
 class DiracKernel;
 
 typedef StoredRange<std::set<const Elem *>::const_iterator, const Elem *> DistElemRange;
@@ -32,7 +32,7 @@ typedef StoredRange<std::set<const Elem *>::const_iterator, const Elem *> DistEl
 class ComputeDiracThread : public ThreadedElementLoop<DistElemRange>
 {
 public:
-  ComputeDiracThread(FEProblem & feproblem, NonlinearSystem & system, SparseMatrix<Number> * jacobian = NULL);
+  ComputeDiracThread(FEProblem & feproblem, NonlinearSystemBase & system, SparseMatrix<Number> * jacobian = NULL);
 
   // Splitting Constructor
   ComputeDiracThread(ComputeDiracThread & x, Threads::split);
@@ -49,7 +49,7 @@ public:
 
 protected:
   SparseMatrix<Number> * _jacobian;
-  NonlinearSystem & _sys;
+  NonlinearSystemBase & _sys;
 
   /// Storage for DiracKernel objects
   const MooseObjectWarehouse<DiracKernel> & _dirac_kernels;
