+++ framework/include/base/ComputeElemDampingThread.h
@@ -23,13 +23,13 @@
 #include "libmesh/elem_range.h"
 
 // Forward declarations
-class NonlinearSystem;
+class NonlinearSystemBase;
 class ElementDamper;
 
 class ComputeElemDampingThread : public ThreadedElementLoop<ConstElemRange>
 {
 public:
-  ComputeElemDampingThread(FEProblem & feproblem, NonlinearSystem & sys);
+  ComputeElemDampingThread(FEProblem & feproblem, NonlinearSystemBase & sys);
 
   // Splitting Constructor
   ComputeElemDampingThread(ComputeElemDampingThread & x, Threads::split split);
@@ -44,7 +44,7 @@ public:
 
 protected:
   Real _damping;
-  NonlinearSystem & _nl;
+  NonlinearSystemBase & _nl;
   const MooseObjectWarehouse<ElementDamper> & _element_dampers;
 };
